Updates migration cmd to change metrics port
This change:

Removes the defaulting in the code where we set DiagnosticsAddress. It
was being overwritten by --diagnostics-address in capiflags.AddManagerOptions() and
setting the default value to :8443.

Updates the deployment manifests to pass --diagnostics-address with a
value of :8081, so we don't clash with the cluster-capi-operator container running in
the pod.