migration: Don't shortcut reconciliation when status is not set
We were assuming that any object without authoritativeAPI in status was
already in the expected migration state. This change removes that
assumption and allows the migration controller to reconcile the object
before marking it as complete. This also ensures that the migration
condition is set correctly, as the shortcut doesn't fully update the
object status.