manifests-gen: set failurePolicy to Ignore for IPAM webhooks
Added functions to set the failurePolicy to Ignore for both mutating and validating webhooks handling IPAM resources. This change addresses issues during bootstrap where the Kube API Server cannot reach webhooks, preventing the creation of IPAddress and IPAddressClaim resources. The new functions ensure that resources can be created even when webhooks are unreachable, aligning with the behavior of the Machine API.